[STREAMPIPES-563] Update data lake sink to StreamPipesDataSink API
